rm submodule structure
