+++ src/abstractinterpretation.jl
@@ -107,8 +107,8 @@ end
 # FIXME: this is such an horrible and super fragile copy-and-paste ...
 # the whole point is to not let abstract interpretation to halt even after there is an
 # obvious error point found (, which is usually represented by `Bottom`-annotated type and
-# triggers early-loop-escape in the native implementation), and allow TP to collect
-# as much errors as possible; so the only diffs are:
+# triggers early-loop-escape in the native implementation), and allow TP to collect as much
+# errors as possible; so the only diffs are:
 # - the first argument changed to `TPInterpreter`
 # - comment out 3 early-loop-escapes
